src/requirements.txt
Flask
python-dotenv
requests
webauthn>=2.0.0
passkeys
user-agents